Readme
Problem Statement:
The data set provided to you is the data set of a Café Chain for one of its restaurants. Do a thorough analysis of the data and come up with the following analysis. The owner of the restaurant wants you to use this data to come up with a set of recommendations that can help his Café Chain increase its revenues. He is able to provide you with a data set for POS (point of sale data) for one of his chains.
	Exploratory Analysis of data & an executive summary (in PPT) of your top findings, supported by graphs. • What kind of trends do you notice in terms of consumer behavior over different times of the day and different days of the week? Can you give concrete recommendations based on the same?  • Are there certain menu items that can be taken off the menu? • Are there trends across months that you are able to notice? 
	Menu Analysis- • Identify the most popular combos that can be suggested to the restaurant chain after a thorough analysis of the most commonly occurring sets of menu items in the customer orders. The restaurant doesn’t have any combo meals. Can you suggest the best combo meals? o Use of Market Basket Analysis (Association Rules)  § Write Something about the association rule and its relevance in this case § Add KNIME workflow Image or Python package used § Write about threshold values of Support and Confidence. Q. Associations Identified § Put the associations in a tabular manner § Explain about support, confidence, & lift values that are calculated. § The suggestion of Possible Combos with Lucrative Offers  § Write recommendations § Make discount offers or combos ( or buy two get one free) based on the associations and your experience. 
Tools Used: Python , Tableau & KNIME
Data Source: Café_Data_MRA.xlsx